pyproject.toml
[project]
name = "hnet-impl"
version = "0.1.0"
description = ""
authors = [
    { email = "187762902+main-horse@users.noreply.github.com" }
]
requires-python = ">=3.11"
dependencies = [
  "torch==2.7.1",
  "mamba-ssm",
  "termcolor>=3.1.0",
  "omegaconf>=2.3.0",
  "h5py>=3.15.1",
  "lightning>=2.6.0",
  "wandb>=0.23.1",
  "nvidia-pytriton==0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[tool.hatch.build.targets.wheel]
packages = ["src/hnet_impl"]

[project.optional-dependencies]
build = [
    "flash-attn==2.8.1",
    "causal-conv1d==1.5.2",
]
[tool.uv]
no-build-isolation-package = ["flash-attn", "causal-conv1d"]
[tool.uv.sources]
mamba-ssm = { git = "https://github.com/main-horse/mamba", rev = "tc" }
[[tool.uv.dependency-metadata]]
name = "flash-attn"
version = "2.8.1"
requires-dist = ["torch"]

[dependency-groups]
dev = [
    "lovely-tensors",
    "pytest>=8.4.1",
    "ruff>=0.12.9",
]


[tool.ruff.lint]
ignore = [
    "E701",  # multiple statements on one line (colon)
    "E741",  # ambiguous variable name
    "F401",  # imported but unused
    "F403",  # 'from module import *' used
    "F821",  # Undefined name `HNetLM`
]
